pyproject.toml
[project]
name = "docuquery-backend"
version = "1.0.0"
description = "RAG-based Smart Document Q&A API"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",

    # Supabase
    "supabase>=2.11.0",
    "python-jose[cryptography]>=3.3.0",

    # OpenAI
    "openai>=1.59.5",
    "tiktoken>=0.8.0",

    # Document Parsing
    "pymupdf>=1.25.1",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",

    # Chunking & Embeddings
    "numpy>=2.2.1",

    # Environment
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.0.0",
]
